Added ability to set the timeout of Raxios via the RaxiosConfig struct

+++ src/raxios_config.rs
@@ -2,6 +2,6 @@ use crate::RaxiosHeaders;
 
 #[derive(Default, Debug, Clone)]
 pub struct RaxiosConfig {
-    pub timeout_ms: Option<usize>,
+    pub timeout_ms: Option<u64>,
     pub headers: Option<RaxiosHeaders>,
 }
